refactor(doctor): compose backend checks from the registry
core/doctor.ts hardcoded the five backend ids and reached into
backend/claude-sdk and backend/codex through dynamic imports for the
model audit and auth detection — the last core→backend edges, kept
visible as the doctor-probes-move-behind-registry warning. The cost was
that adding a backend did not add its checks unless doctor was taught
about it.

BackendFactory gains an optional doctor(config, isActive) slot. Each
built-in supplies its own probes, moved verbatim from doctor.ts:

- claude-sdk/doctor.ts — the CLI binary and, for the active backend
  only, the configured-model audit against the static catalog.
- codex/doctor.ts — the CLI binary and detectCodexAuth.
- openai-agents/doctor.ts — API key source and endpoint.
- remote-server/factory.ts — the "SDK spawns a CLI of the same name"
  binary check, once for Kilo and OpenCode.

doctor.ts composes whatever is registered: the active backend is
checked in full, every other registered (or config-exposed) backend is
reported as inactive with failures downgraded, an active backend that
is not registered is a fail that names the known ids, and a factory
without the slot is an info line rather than a silent skip. Errors
thrown by a probe become a warn check instead of killing the report.

The types doctor and the registry share (DoctorCheck, DoctorConfigSlice,
DoctorReport, NativeModuleCheck) move to core/doctor-types.ts so the
registry can declare the slot without importing the collector; doctor.ts
re-exports them. binaryOnPath moves to util/binary-on-path.ts since three
backends now need it.

Loading: backend/builtins.ts is the one list of factories, used by
bootstrap, by `talon doctor` (which runs standalone, so nothing else has
registered them), and by doctor.test.ts. The depcruise rule ratchets to
error — 594 modules, 0 errors, and the in-flight warnings drop 18 → 13.
